Fix daily jobs crash: enum name in SQL and missing error isolation (#2883)
StatGenerator crashed with "Invalid column name 'New'" because
LitterReportStatusEnum.New was string-concatenated as its name
instead of its integer value. Since Program.cs had no error
isolation, this unhandled exception prevented LeaderboardGenerator
and AchievementProcessor from running — causing stale leaderboards
and achievements in production.

Fixes #2881

+++ TrashMobDailyJobs/StatGenerator.cs
@@ -218,7 +218,7 @@ namespace TrashMobDailyJobs
 
         private async Task<int> CountNewLitterReports(SqlConnection conn)
         {
-            var sql = "SELECT count(*) FROM dbo.LitterReports where LitterReportStatusId = " + LitterReportStatusEnum.New;
+            var sql = "SELECT count(*) FROM dbo.LitterReports where LitterReportStatusId = " + (int)LitterReportStatusEnum.New;
             var numberOfLitterReports = 0;
 
             using (var cmd = new SqlCommand(sql, conn))
@@ -234,7 +234,7 @@ namespace TrashMobDailyJobs
 
         private async Task<int> CountCleanedLitterReports(SqlConnection conn)
         {
-            var sql = "SELECT count(*) FROM dbo.LitterReports where LitterReportStatusId = " + LitterReportStatusEnum.Cleaned;
+            var sql = "SELECT count(*) FROM dbo.LitterReports where LitterReportStatusId = " + (int)LitterReportStatusEnum.Cleaned;
             var numberOfLitterReports = 0;
 
             using (var cmd = new SqlCommand(sql, conn))
